PIN: move setting to a dedicated screen (no other change)
