+++ vector/src/main/res/xml/vector_settings_pin.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.preference.PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <im.vector.app.core.preference.VectorSwitchPreference
+        android:defaultValue="false"
+        android:key="SETTINGS_SECURITY_USE_PIN_CODE_FLAG"
+        android:summary="@string/settings_security_pin_code_summary"
+        android:title="@string/settings_security_pin_code_title" />
+
+</androidx.preference.PreferenceScreen>
